src/test/java/year2023/day07/Day07Test.java
package year2023.day07;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

import java.util.Arrays;
import java.util.stream.Collectors;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class Day07Test {

    @Test
    void testSolveA() {
        String input = """
                32T3K 765
                T55J5 684
                KK677 28
                KTJJT 220
                QQQJA 483
                """;
        var inputs = Arrays.stream(input.split("\n"))
                .filter(s->!s.isEmpty())
                .map(String::valueOf)
                .collect(Collectors.toList());
        Assertions.assertEquals(6440L, new Day07().solveA(inputs));
    }
    @Test
    void testSolveB() {
        String input = """
                KTJJT 220
                32T3K 765
                T55J5 684
                KK677 28
                QQQJA 483
                """;
        var inputs = Arrays.stream(input.split("\n"))
                .filter(s->!s.isEmpty())
                .map(String::valueOf)
                .collect(Collectors.toList());
        assertEquals(5905L, new Day07().solveB(inputs));
    }
}
